feat: add ability to Expand/Collapse All
This introduces a new property named `showExpandAll` (defaults to `false`)
that display a button that allows to easily expand or collapse all rows.

The button shows up in both the breadcrumbs toolbar as above the top schema row.

Replace `jotai` related deprecated functions to the suggested new alternative
functions to avoid the spamming of deprecation warnings in the developer console.